+++ @xen-orchestra/lite/src/views/new-vm/NewVmView.vue
@@ -157,7 +157,16 @@
               <UiInput v-model="vmState.vCPU" accent="brand" :label="$t('vcpus', Number(vmState.vCPU))" />
               <!-- TODO remove (GB) when we can use new selector -->
               <UiInput v-model="ramFormatted" accent="brand" :label="`${$t('ram')} (GB)`" />
-              <UiInput v-model="vmState.topology" accent="brand" disabled :label="$t('topology')" />
+              <VtsInputWrapper :label="$t('topology')">
+                <div class="topology">
+                  <FormSelect v-model="vmState.topology">
+                    <option :value="null">{{ $t('default-behavior') }}</option>
+                    <option v-for="core in coresPerSocket" :key="core.value" :value="core.value">
+                      {{ core.label }}
+                    </option>
+                  </FormSelect>
+                </div>
+              </VtsInputWrapper>
             </div>
             <!-- NETWORK SECTION -->
             <UiTitle>{{ $t('network') }}</UiTitle>
@@ -472,7 +481,7 @@ const vmState = reactive<VmState>({
   vCPU: 0,
   VCPUs_max: 0,
   ram: 0,
-  topology: '',
+  topology: null,
   copyHostBiosStrings: false,
   sshKeys: [],
   existingVdis: [],
@@ -528,6 +537,40 @@ const deleteItem = <T,>(array: T[], index: number) => {
 }
 
 const poolName = computed(() => pool.value?.name_label)
+
+const poolCpuInfo = computed(() => {
+  return {
+    cpus: {
+      cores: pool.value?.cpu_info && +pool.value.cpu_info.cpu_count,
+      sockets: pool.value?.cpu_info && +pool.value.cpu_info.socket_count,
+    },
+  }
+})
+
+const coresPerSocket = computed(() => {
+  // https://github.com/xcp-ng/xenadmin/blob/0160cd0119fae3b871eef656c23e2b76fcc04cb5/XenModel/XenAPI-Extensions/VM.cs#L62
+  const MAX_VM_SOCKETS = 16
+  const minCores = vmState.vCPU / MAX_VM_SOCKETS
+  const options = []
+  let cores = poolCpuInfo.value.cpus.cores
+
+  if (cores === undefined || vmState.vCPU === undefined) return []
+
+  for (cores; cores >= minCores; cores--) {
+    if (vmState.vCPU % cores === 0) {
+      options.push({
+        // TODO Need to improve pluralization
+        label: t('vmSocketsWithCoresPerSocket', {
+          nSockets: vmState.vCPU / cores,
+          nCores: cores,
+        }),
+        value: cores,
+      })
+    }
+  }
+
+  return options
+})
 
 const hostMasterRef = computed(() => pool.value?.master)
 
@@ -712,6 +755,7 @@ const vmCreationParams = computed(() => ({
   bootAfterCreate: vmState.boot_vm,
   copyHostBiosStrings: vmState.boot_firmware !== 'uefi' && !templateHasBiosStrings.value && vmState.copyHostBiosStrings,
   hvmBootFirmware: vmState.boot_firmware,
+  coresPerSocket: vmState.topology,
   tags: vmState.tags,
   cloudConfig: '',
 }))
@@ -764,9 +808,10 @@ const _createVm = async ($defer: Defer) => {
       xapi.vm.setNameLabel(vmRefs, vmCreationParams.value.name_label),
       xapi.vm.setNameDescription(vmRefs, vmCreationParams.value.name_description),
       xapi.vm.setMemory(vmRefs, vmCreationParams.value.memory),
-      xapi.vm.setVCPUsAtStartup(vmRefs, vmCreationParams.value.cpus),
-      xapi.vm.setHvmBootFirmware(vmRefs[0], vmCreationParams.value.hvmBootFirmware),
       xapi.vm.setAutoPowerOn(vmRefs[0], vmCreationParams.value.autoPoweron),
+      xapi.vm.setCorePerSocket(vmRefs[0], vmCreationParams.value.coresPerSocket),
+      xapi.vm.setHvmBootFirmware(vmRefs[0], vmCreationParams.value.hvmBootFirmware),
+      xapi.vm.setVCPUsAtStartup(vmRefs, vmCreationParams.value.cpus),
     ])
 
     // INSTALL SETTINGS
@@ -979,6 +1024,10 @@ const createVM = defer(_createVm)
     .memory-container {
       display: flex;
       gap: 10.8rem;
+
+      .topology {
+        width: 32rem;
+      }
     }
 
     thead tr th:last-child {